feat: 背景透過してベストスコアのダイアログを表示できるよう修正しました
